ML/teaching/llm/requirements.txt
python-dotenv
openai
mcp
langchain
langchain-openai
